requirements.txt
langchain
openai
pinecone-client
tiktoken
nest_asyncio